Refactor aggregate persistence (#1)
* Remove EventSourcedAggregateRoot
* Fix unit-tests
* Update Relational repo
* Refactor AggregateVersion
* Refactor entity Id handling
* Move EnsureConsistency to AggregaRoot baseclass
* Make TryHandleDomainEvent virtual
* Remove obsolete code
* Add Document logic
* Move generic persistence code to infra.common

+++ src/Common/Domain.Common/Aggregate/Command.cs
@@ -15,7 +15,7 @@ public record Command
     /// The expected version of the aggregate to apply the command to. This is 
     /// used for optimistic concurrency control.
     /// </summary>
-    public ulong? ExpectedVersion { get; set; }
+    public uint? ExpectedVersion { get; set; }
 
     /// <summary>
     /// The type of the command. This value is primarily used for logging and
